Correccion: cambio de mapeo PlayerEntity -> TournamentParticipantsEntity
